Moving the function that calculates and queries allocation setup
